Folder that contains the scripts used for WCSim as well as some WCSim…
output files. reflectorScript.sh runs WCSim with a reflector of given dimensions and offsets. Instructions on how to set the parameters are located at the top of the script as comments. reflectorScript.sh sends the jobs to the neut cluster, the cluster scripts are created by reflectorScript.sh and stored in the clusterScripts directory. Log files from the run are stored in the logs directory. Initial run output is stored in neutRun1. cedarFiles are taken from a run on cedar and moved here for analysis purposes. Analysis script (cedarFiles/plot_tp.cc) is an slightly modified analysis script that creates a 2D Graph with the files in that folder. To use this script, simply copy it to a different location and change line 54 to represent the filepath that it was moved to.